Remove JNA-specific uses in it-domino-jnx
